// packages/ng-platform/src/api/common.model.ts
/*
 * Copyright (c) 2017 Michael Krotscheck
 *
 * Licensed under the Apache License, Version 2.0 (the "License"); you may not
 * use this file except in compliance with the License. You may obtain a copy
 * of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
 * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 *
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

/**
 * This interface describes an application instance.
 *
 * @author Michael Krotscheck
 */
export interface CommonModel {

  /**
   * The entity ID, usually a UUID.
   */
  id?: string;

  /**
   * Created date timestamp, in milliseconds.
   */
  createdDate?: number;

  /**
   * Last modified date timestamp, in milliseconds.
   */
  modifiedDate?: number;
}